pyproject.toml
[project]
name = "multimodal-agentic-app-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-huggingface>=1.2.2",
]

[tool.uv.workspace]
members = [
    ".",
]

[tool.uv.sources]
multimodal-agentic-app-rag = { workspace = true }

[dependency-groups]
dev = [
    "flask>=3.1.3",
    "langchain>=1.3.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.2",
    "langchain-groq>=1.1.2",
    "langchain-pinecone>=0.2.13",
    "langchain-tavily>=0.2.18",
    "langchain-text-splitters>=1.1.2",
    "langgraph>=1.2.0",
    "multimodal-agentic-app-rag",
    "pinecone-client>=6.0.0",
    "pypdf>=6.11.0",
    "sentence-transformers>=5.5.0",
]
